Add files via upload
SFML setup and one button now. Have to give functionality to switch to the second screen. The plan, for now, is to allow the user to search for a car with a DFS vs. a BFS and compare the time it takes to search for the car. Also added arial font to put text on sfml.